# api/docs/swagger.yaml
definitions:
...
  response.body:
    properties:
      data:
        description: payload data (on success)
      errors:
        description: detailed error list
        items:
          type: string
        type: array
      message:
        description: short message (on failure)
        type: string
      success:
        description: 'status: success or failure'
        type: boolean
    type: object
  user.CreateUserRequest:
    properties:
      email:
        description: Required, valid email format
        example: john.doe@example.com
        type: string
      name:
        description: Required, 3–100 characters
        example: John Doe
        maxLength: 100
        minLength: 3
        type: string
      password:
        description: Required, 8–72 characters (72 = bcrypt limit)
        example: password
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  user.UserResponse:
    properties:
      created_at:
        description: Creation time (RFC3339 format)
        type: string
      email:
        description: User email
        type: string
      id:
        description: User ID
        type: integer
      name:
        description: User name
        type: string
    type: object
info:
  contact: {}
  description: Pepeng Book Journal and Notes API
  title: PengBook API
  version: "1.0"
paths:
...
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: Creates a new user with name, email and password
      parameters:
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.body'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.body'
            - properties:
                errors:
                  items:
                    type: string
                  type: array
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.body'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      summary: Create a new user
      tags:
      - users
  /api/v1/users/{id}:
    get:
      description: Fetch a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.body'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
                success:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.body'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.body'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.body'
            - properties:
                message:
                  type: string
                success:
                  type: boolean
              type: object
      summary: Get user by ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"